extract RejectIcUserRegistrationToAgencyForm in a new file
